extended the "Language" custom taxonomy by adding media uploader
Copied the template code from here:
https://gist.github.com/mathetos/1eea92f71934442671a7

So far, I've got it to work so that user can add/edit/remove images
attached to a Language. Also quickly tested this by retrieving image
source in the index.php and it works.

Need to refactor so that the code is more re-usable. Also, need to
change how it looks a little, and may need to move it into another .php
file