Fix for "Set Featured Image doesnt work" #9
It seems the issue is that in the industries forms, the wp core
functionality "Feature Image" is clashing with the "MultiPostThumbnails"
functionality. We still need 2 image controls in the industries form so
what I have done is removed the wp core functionality "feature Image"
and instead added a second "MultiPostThumbnail" control to act as the
Feature Image. I've had to change the way the feature image is being
retrieved in the home page as well